Stepper code refinment, Initial code porting to MP lab
The Stepper code has been experiment with to get better results this has
had little effect. The results where put into an excell file to easier
interprit them.

The PIC2 4FJ128GA202 has been chosen and ordered for this project. It
was used as the basis for the simulation in MP lab. Code seems to
compile and used debugger the same step count was measured as when
compiling with MinGW